requirements.txt
requests
python-dotenv>=1.0.0